% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vsmi.R
\docType{package}
\name{vsmi}
\alias{vsmi-package}
\alias{vsmi}
\title{vsmi: Variable selection for multiple imputed data}
\description{
This is a package to implementation
penalized weighted least-squares estimate for variable selection on correlated multiply imputed data
and penalized estimating equations for generalized linear models with multiple imputation.
}
\section{Functions}{

\code{\link[vsmi]{PEE}}:Penalized estimating equations for generalized linear models with multiple imputation

\code{\link[vsmi]{PWLS}} : Penalized weighted least-squares estimate for variable selection on correlated multiply imputed data

\code{\link[vsmi]{generate_pwls_missing_data}} : Generate example missing data for  PWLS

\code{\link[vsmi]{generate_pee_missing_data}} : Generate example missing data for PEE
}

\author{
\strong{Maintainer}: Haoyu Yang \email{haoyuyang@alu.ruc.edu.cn}

Authors:
\itemize{
  \item Mingyue Zhang
  \item Yang Li
}

}
